Change votes to use double.
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4320 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ proctor/include/pcl/proctor/proposer.h
@@ -15,7 +15,7 @@ namespace pcl
 
     struct Candidate {
       std::string id;
-      float votes;
+      double votes;
     };
 
     class Proposer {
